+++ main.go
@@ -66,6 +66,7 @@ func init() {
 func main() {
 	var enableLeaderElection bool
 	var enableLogging bool
+	var loggingAgent string
 	var installationName string
 	var insecureCA bool
 	var metricsAddr string
@@ -76,6 +77,7 @@ func main() {
 		"Enable leader election for controller manager. "+
 			"Enabling this will ensure there is only one active controller manager.")
 	flag.BoolVar(&enableLogging, "enable-logging", true, "enable/disable logging for the whole installation")
+	flag.StringVar(&loggingAgent, "logging-agent", "promtail", "select logging agent to use (promtail or alloy-logs)")
 	flag.StringVar(&installationName, "installation-name", "unknown", "Name of the installation")
 	flag.BoolVar(&insecureCA, "insecure-ca", false, "Is the management cluter CA insecure?")
 	flag.StringVar(&metricsAddr, "metrics-bind-address", ":8080", "The address the metric endpoint binds to.")
@@ -155,6 +157,7 @@ func main() {
 	}
 
 	loggedcluster.O.EnableLoggingFlag = enableLogging
+	loggedcluster.O.LoggingAgent = loggingAgent
 	loggedcluster.O.InstallationName = installationName
 	loggedcluster.O.InsecureCA = insecureCA
 	setupLog.Info("Loggedcluster config", "options", loggedcluster.O)