+++ pkg/logged-cluster/capicluster/cluster.go
@@ -38,6 +38,14 @@ func (o Object) HasLoggingEnabled() bool {
 	return loggingEnabled
 }
 
+func (o Object) GetLoggingAgent() string {
+	return o.Options.LoggingAgent
+}
+
+func (o *Object) SetLoggingAgent(loggingAgent string) {
+	o.Options.LoggingAgent = loggingAgent
+}
+
 func (o Object) IsInsecureCA() bool {
 	return o.Options.InsecureCA
 }
@@ -76,7 +84,7 @@ func (o Object) GetObservabilityBundleConfigMap() string {
 }
 
 func (o Object) getWiredExtraConfig() appv1.AppExtraConfig {
-	observabilityBundleConfigMapMeta := common.ObservabilityBundleConfigMapMeta(o)
+	observabilityBundleConfigMapMeta := common.ObservabilityBundleConfigMapMeta(&o)
 	return appv1.AppExtraConfig{
 		Kind:      "configMap",
 		Name:      observabilityBundleConfigMapMeta.GetName(),