+++ pkg/logged-cluster/vintagewc/cluster.go
@@ -44,6 +44,14 @@ func (o Object) HasLoggingEnabled() bool {
 	return loggingEnabled
 }
 
+func (o Object) GetLoggingAgent() string {
+	return o.Options.LoggingAgent
+}
+
+func (o *Object) SetLoggingAgent(loggingAgent string) {
+	o.Options.LoggingAgent = loggingAgent
+}
+
 func (o Object) IsInsecureCA() bool {
 	return o.Options.InsecureCA
 }