+++ pkg/resource/logging-secret/logging-secret.go
@@ -3,64 +3,50 @@ package loggingsecret
 import (
 	"fmt"
 
-	"github.com/pkg/errors"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/yaml"
+
+	"github.com/pkg/errors"
 
 	"github.com/giantswarm/logging-operator/pkg/common"
 	loggedcluster "github.com/giantswarm/logging-operator/pkg/logged-cluster"
-	loggingcredentials "github.com/giantswarm/logging-operator/pkg/resource/logging-credentials"
 )
 
 const (
 	loggingClientSecretName = "logging-secret"
 )
 
-type values struct {
-	Promtail promtail `yaml:"promtail" json:"promtail"`
-}
+func GenerateLoggingSecret(lc loggedcluster.Interface, loggingCredentialsSecret *v1.Secret, lokiURL string) (v1.Secret, error) {
+	var data map[string][]byte
+	var err error
 
-type promtail struct {
-	Config promtailConfig `yaml:"config" json:"config"`
-}
+	switch lc.GetLoggingAgent() {
+	case "promtail":
+		data, err = GeneratePromtailLoggingSecret(lc, loggingCredentialsSecret, lokiURL)
+		if err != nil {
+			return v1.Secret{}, err
+		}
+	case "alloy-logs":
+		data, err = GenerateAlloyLoggingSecret(lc, loggingCredentialsSecret, lokiURL)
+		if err != nil {
+			return v1.Secret{}, err
+		}
+	default:
+		return v1.Secret{}, errors.Errorf("unsupported logging agent %q", lc.GetLoggingAgent())
+	}
 
-type promtailConfig struct {
-	Clients []promtailConfigClient `yaml:"clients" json:"clients"`
-}
+	secret := v1.Secret{
+		ObjectMeta: SecretMeta(lc),
+		Data:       data,
+	}
 
-// TODO: use upstream promtail structures
-type promtailConfigClient struct {
-	URL            string                             `yaml:"url" json:"url"`
-	TenantID       string                             `yaml:"tenant_id" json:"tenant_id"`
-	BasicAuth      promtailConfigClientBasicAuth      `yaml:"basic_auth" json:"basic_auth"`
-	BackoffConfig  promtailConfigClientBackoffConfig  `yaml:"backoff_config" json:"backoff_config"`
-	ExternalLabels promtailConfigClientExternalLabels `yaml:"external_labels" json:"external_labels"`
-	TLSConfig      promtailConfigClientTLSConfig      `yaml:"tls_config" json:"tls_config"`
-}
-
-type promtailConfigClientTLSConfig struct {
-	InsecureSkipVerify bool `yaml:"insecure_skip_verify" json:"insecure_skip_verify"`
-}
-
-type promtailConfigClientExternalLabels struct {
-	Installation string `yaml:"installation" json:"installation"`
-	ClusterID    string `yaml:"cluster_id" json:"cluster_id"`
-}
-
-type promtailConfigClientBackoffConfig struct {
-	MaxPeriod string `yaml:"max_period" json:"max_period"`
-}
-
-type promtailConfigClientBasicAuth struct {
-	Username string `yaml:"username" json:"username"`
-	Password string `yaml:"password" json:"password"`
+	return secret, nil
 }
 
 // SecretMeta returns metadata for the logging-secret
 func SecretMeta(lc loggedcluster.Interface) metav1.ObjectMeta {
 	metadata := metav1.ObjectMeta{
-		Name:      getLoggingSecretName(lc),
+		Name:      GetLoggingSecretName(lc),
 		Namespace: lc.GetAppsNamespace(),
 		Labels:    map[string]string{},
 	}
@@ -69,62 +55,6 @@ func SecretMeta(lc loggedcluster.Interface) metav1.ObjectMeta {
 	return metadata
 }
 
-// GenerateLoggingSecret returns a secret for
-// the Loki-multi-tenant-proxy auth config
-func GenerateLoggingSecret(lc loggedcluster.Interface, credentialsSecret *v1.Secret, lokiURL string) (v1.Secret, error) {
-	clusterName := lc.GetClusterName()
-
-	writeUser := clusterName
-
-	writePassword, err := loggingcredentials.GetPassword(lc, credentialsSecret, clusterName)
-	if err != nil {
-		return v1.Secret{}, errors.WithStack(err)
-	}
-
-	installationName := lc.GetInstallationName()
-
-	values := values{
-		Promtail: promtail{
-			Config: promtailConfig{
-				Clients: []promtailConfigClient{
-					{
-						URL:      fmt.Sprintf("https://%s/loki/api/v1/push", lokiURL),
-						TenantID: clusterName,
-						BasicAuth: promtailConfigClientBasicAuth{
-							Username: writeUser,
-							Password: writePassword,
-						},
-						BackoffConfig: promtailConfigClientBackoffConfig{
-							MaxPeriod: "10m",
-						},
-						ExternalLabels: promtailConfigClientExternalLabels{
-							Installation: installationName,
-							ClusterID:    lc.GetClusterName(),
-						},
-						TLSConfig: promtailConfigClientTLSConfig{
-							InsecureSkipVerify: lc.IsInsecureCA(),
-						},
-					},
-				},
-			},
-		},
-	}
-
-	v, err := yaml.Marshal(values)
-	if err != nil {
-		return v1.Secret{}, errors.WithStack(err)
-	}
-
-	secret := v1.Secret{
-		ObjectMeta: SecretMeta(lc),
-		Data: map[string][]byte{
-			"values": []byte(v),
-		},
-	}
-
-	return secret, nil
-}
-
-func getLoggingSecretName(lc loggedcluster.Interface) string {
+func GetLoggingSecretName(lc loggedcluster.Interface) string {
 	return fmt.Sprintf("%s-%s", lc.GetClusterName(), loggingClientSecretName)
 }